Add stable/2402 branch to docker executor image scripts
- Refactor csit_pip_cache function to use csit bash
  functions that activate the python virtualenv in
  production jobs.
- Implement workaround for CSIT python cache miss
  apparently due to change of cache directory in
  pip 23.x and newer.

+++ docker/scripts/lib_csit.sh
@@ -69,14 +69,14 @@ csit_install_packages() {
     git clean -qfdx
     python3 -m pip install pyyaml
 
-    local exclude_roles="-e calibration -e kernel -e mellanox -e nomad -e consul"
+    local exclude_roles="-e calibration -e kernel -e mellanox -e nomad -e consul -e aws -e vpp"
     [ "$OS_ARCH" = "aarch64" ] && exclude_roles="$exclude_roles -e iperf"
 
     # Not in double quotes to let bash remove newline characters
     local yaml_files
     yaml_files="$(grep -r packages_by $csit_ansible_dir | cut -d: -f1 | sort -u | grep -v $exclude_roles)"
     packages="$(dbld_csit_find_ansible_packages.py --$OS_ID --$OS_ARCH $yaml_files)"
-    packages="${packages/bionic /}"
+    packages="${packages/jammy /}"
     packages="${packages/focal /}"
     packages="${packages/libmbedcrypto1/libmbedcrypto3}"
     packages="${packages/libmbedtls10/libmbedtls12}"
@@ -97,7 +97,35 @@ csit_install_packages() {
     fi
 }
 
-csit_pip_cache() {
+csit_install_hugo() {
+    local branch="$1"
+    CSIT_DIR="$DOCKER_CSIT_DIR"
+
+    if [ -f "$CSIT_DIR/VPP_REPO_URL" ] \
+           && [ -f "$CSIT_DIR/requirements.txt" ]; then
+
+        local branchname
+        # use bash variable substitution to replace '/' with '_' to convert from
+        # vpp to csit branch name nomenclature
+        branchname="${branch////_}"
+        local csit_bash_function_dir="$CSIT_DIR/resources/libraries/bash/function"
+        local bld_log="$DOCKER_BUILD_LOG_DIR"
+        bld_log="${bld_log}/$FDIOTOOLS_IMAGENAME-$branchname-csit_pip_cache-bld.log"
+
+        description="Install CSIT hugo packages from $branch branch"
+        echo_log "    Starting  $description..."
+        git clean -qfdx
+
+        source $csit_bash_function_dir/hugo.sh
+        go_install
+        hugo_install
+    else
+        echo_log "ERROR: Missing or invalid CSIT_DIR: '$CSIT_DIR'!"
+        return 1
+    fi
+}
+
+to_be_deprecated_csit_pip_cache() {
     local branch="$1"
     # ensure PS1 is defined (used by virtualenv activate script)
     PS1=${PS1:-"#"}
@@ -150,20 +178,86 @@ csit_pip_cache() {
 
             rm -f "$constraintfile"
         fi
-
         # Install csit python requirements
         $pip_cmd install -r "$CSIT_DIR/requirements.txt" 2>&1 | \
             tee -a "$bld_log"
         # Install tox python requirements
         $pip_cmd install -r "$CSIT_DIR/tox-requirements.txt" 2>&1 | \
             tee -a "$bld_log"
+        $pip_cmd install pip --upgrade
         # Run tox which installs pylint requirments
         pushd "$CSIT_DIR" >& /dev/null || exit 1
-        tox || true
+        tox -e pylint || true
         popd >& /dev/null || exit 1
 
         # Clean up virtualenv directories
         deactivate
+        # Install csit python requirements
+        $pip_cmd install -r "$CSIT_DIR/requirements.txt" 2>&1 | \
+            tee -a "$bld_log"
+        # Install tox python requirements
+        $pip_cmd install -r "$CSIT_DIR/tox-requirements.txt" 2>&1 | \
+            tee -a "$bld_log"
+        git checkout -q -- .
+        git clean -qfdx
+        echo_log "    Completed $description!"
+    else
+        echo_log "ERROR: Missing or invalid CSIT_DIR: '$CSIT_DIR'!"
+        return 1
+    fi
+}
+
+csit_pip_cache() {
+    local branch="$1"
+    # ensure PS1 is defined (used by virtualenv activate script)
+    PS1=${PS1:-"#"}
+    CSIT_DIR="$DOCKER_CSIT_DIR"
+
+    if [ -f "$CSIT_DIR/VPP_REPO_URL" ] \
+           && [ -f "$CSIT_DIR/requirements.txt" ]; then
+
+        local branchname
+        # use bash variable substitution to replace '/' with '_' to convert from
+        # vpp to csit branch name nomenclature
+        branchname="${branch////_}"
+        local csit_bash_function_dir="$CSIT_DIR/resources/libraries/bash/function"
+        local bld_log="$DOCKER_BUILD_LOG_DIR"
+        bld_log="${bld_log}/$FDIOTOOLS_IMAGENAME-$branchname-csit_pip_cache-bld.log"
+        local pip_cmd="python3 -m pip --disable-pip-version-check"
+        export PYTHONPATH=$CSIT_DIR
+
+        description="Install CSIT python packages from $branch branch"
+        echo_log "    Starting  $description..."
+        git clean -qfdx
+        rm -rf "$PYTHONPATH/env"
+
+        # Activate / install CSIT python virtualenv ($CSIT_DIR/requirements.txt)
+        local common_sh="$csit_bash_function_dir/common.sh"
+        source $common_sh
+        # TODO: figure out what happened to virtualenv/pip that causes a activate
+        # to fail to instantiate the venv & a cache SNAFU that causes cache
+        # misses in the CI.  Running the activate_virtualenv function twice populates
+        # the appropriate pip cache (.cache/pip/http dir was moved to http-v2 in
+        # pip3 23.x) appropriately so do it twice.
+        activate_virtualenv "${CSIT_DIR}" "${CSIT_DIR}/requirements.txt" | tee -a "$bld_log"
+        command -v "deactivate" || \
+            activate_virtualenv "${CSIT_DIR}" "${CSIT_DIR}/requirements.txt" | tee -a "$bld_log"
+        command -v "deactivate" && deactivate
+
+        # Install tox python requirements
+        activate_virtualenv "${CSIT_DIR}" "${CSIT_DIR}/tox-requirements.txt" |\
+            tee -a "$bld_log"
+        command -v "deactivate" || \
+            activate_virtualenv "${CSIT_DIR}" "${CSIT_DIR}/tox-requirements.txt" |\
+            tee -a "$bld_log"
+
+        # Run tox which installs pylint requirements
+        pushd "$CSIT_DIR" >& /dev/null || exit 1
+        tox -e pylint || true
+        popd >& /dev/null || exit 1
+
+        # Clean up virtualenv directories
+        command -v "deactivate" && deactivate
         git checkout -q -- .
         git clean -qfdx
         echo_log "    Completed $description!"
